# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Politics Project",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/classification_results": {
      "post": {
        "tags": [
          "classification-results",
          "ClassificationResults"
        ],
        "summary": "Create Classification Result",
        "description": "Create (store) an individual classification result.\nVerifies that the workspace exists and that the referenced document and scheme belong to that workspace.",
        "operationId": "create_classification_result",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationResultCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationResultRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "classification-results",
          "ClassificationResults"
        ],
        "summary": "List Classification Results",
        "description": "List all classification results for the given workspace.",
        "operationId": "list_classification_results",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          },
          {
            "name": "document_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "title": "Document Ids"
            }
          },
          {
            "name": "scheme_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "title": "Scheme Ids"
            }
          },
          {
            "name": "run_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Name"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnhancedClassificationResultRead"
                  },
                  "title": "Response Classification-Results-List Classification Results"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/classification_results/{result_id}": {
      "get": {
        "tags": [
          "classification-results",
          "ClassificationResults"
        ],
        "summary": "Get Classification Result",
        "description": "Load (retrieve) an individual classification result by its ID.\nVerifies that this result's document and scheme belong to the workspace.",
        "operationId": "get_classification_result",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          },
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationResultRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/classification_results/": {
      "get": {
        "tags": [
          "classification-results",
          "ClassificationResults"
        ],
        "summary": "List Classification Results",
        "description": "List all classification results for the given workspace.",
        "operationId": "list_classification_results",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          },
          {
            "name": "document_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "title": "Document Ids"
            }
          },
          {
            "name": "scheme_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "title": "Scheme Ids"
            }
          },
          {
            "name": "run_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Name"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnhancedClassificationResultRead"
                  },
                  "title": "Response Classification-Results-List Classification Results"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassificationResultCreate": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "scheme_id": {
            "type": "integer",
            "title": "Scheme Id"
          },
          "value": {
            "type": "object",
            "title": "Value"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name"
          },
          "run_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Description"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "scheme_id"
        ],
        "title": "ClassificationResultCreate"
      },
      "ClassificationResultRead": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "scheme_id": {
            "type": "integer",
            "title": "Scheme Id"
          },
          "value": {
            "type": "object",
            "title": "Value"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name"
          },
          "run_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document": {
            "$ref": "#/components/schemas/DocumentRead"
          },
          "scheme": {
            "$ref": "#/components/schemas/ClassificationSchemeRead"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "scheme_id",
          "timestamp",
          "id",
          "document",
          "scheme"
        ],
        "title": "ClassificationResultRead"
      },
...
      "ClassificationSchemeRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Type of classification scheme"
          },
          "scale_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Min",
            "default": 0
          },
          "scale_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Max",
            "default": 10
          },
          "is_set_of_labels": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Set Of Labels",
            "default": false
          },
          "max_labels": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Labels"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "dict_keys": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dict Keys"
          },
          "model_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Instructions"
          },
          "validation_rules": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Rules"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "classification_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Count"
          },
          "document_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "type",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "ClassificationSchemeRead"
      },
...
      "DocumentRead": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type",
            "default": "article"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "top_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Image"
          },
          "text_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "insertion_date": {
            "type": "string",
            "format": "date-time",
            "title": "Insertion Date"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileRead"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "insertion_date",
          "id",
          "workspace_id",
          "user_id"
        ],
        "title": "DocumentRead"
      },
...
      "EnhancedClassificationResultRead": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "scheme_id": {
            "type": "integer",
            "title": "Scheme Id"
          },
          "value": {
            "type": "object",
            "title": "Value"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name"
          },
          "run_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document": {
            "$ref": "#/components/schemas/DocumentRead"
          },
          "scheme": {
            "$ref": "#/components/schemas/ClassificationSchemeRead"
          },
          "display_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Value"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "scheme_id",
          "timestamp",
          "id",
          "document",
          "scheme"
        ],
        "title": "EnhancedClassificationResultRead"
      },
      "FileRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "filetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filetype"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "top_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Image"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "document_id"
        ],
        "title": "FileRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}